Project Done with Frontend and Backend Working

+++ backend/controller/product.controller.js
@@ -8,13 +8,13 @@ export const getProducts = async (req, res) => {
         res.status(200).json({success: true, data: products}); 
     }
     catch (error) { 
-        console.log("Error in fetching all of the datas"); 
+        console.log("Error in fetching all of the datas", error.message); 
         res.status(500).json({success: false, message: "Server error"}); 
     }
 }
 
 // 2nd Post Products controller
-export const postProducts = async (req, res) => { 
+export const createProducts = async (req, res) => { 
     const product = req.body; // This is when the user will send the data
     
     if (!product.name || !product.price || !product.image) { 
@@ -22,9 +22,8 @@ export const postProducts = async (req, res) => {
     }
     
     // Assuming the data we got is working, in the sense that this product is valid, I want it to be in the format of the schema that we made
-    const newProduct = new Product(product); 
-        
     try { 
+        const newProduct = new Product(product); 
         await newProduct.save(); // Wait till the newProduct is saved in the database
         res.status(201).json({success: true, data: newProduct}); 
     }
@@ -34,22 +33,19 @@ export const postProducts = async (req, res) => {
     }
 }
 
-export const putProducts = async (req, res) => {
-    const product = req.body; // This is when the user will send the data
+export const updateProducts = async (req, res) => {
+    const { id } = req.params; 
+    const product = req.body; 
 
-    if (!product.name || !product.price || !product.image) { 
-        return res.status(400).json({success: false, message: "Please provide all fields"}); 
+    if (!mongoose.Types.ObjectId.isValid(id)) { 
+        return res.status(404).json({success: false, message: "Invalid prodcut id"}); 
     }
 
-    // Assuming the data we got is working, in the sense that this product is valid, I want it to be in the format of the schema that we made
-    const newProduct = new Product(product); 
-    
     try { 
-        await newProduct.save(); // Wait till the newProduct is saved in the database
-        res.status(201).json({success: true, data: newProduct()}); 
+        const updatedProduct = await Product.findByIdAndUpdate(id, product, {new: true}); 
+        res.status(200).json({success: true, data: updatedProduct}); 
     }
-    catch(error) { 
-        console.log("Error in creating a product:", error.message); 
+    catch (error) { 
         res.status(500).json({success: false, message: "Server Error"}); 
     }
 }
@@ -57,11 +53,16 @@ export const putProducts = async (req, res) => {
 export const deleteProducts = async (req, res) => { 
     const {id} = req.params; 
 
+    if (!mongoose.Types.ObjectId.isValid(id)) { 
+        return res.status(404).json({success: false, message: "Invalid product id"}); 
+    }
+
     try { 
         await Product.findByIdAndDelete(id); 
         res.status(200).json({success: true, message: "product is deleted"}); 
     }
     catch (error) { 
+        console.log("Error in deleting product:", error.message);
         res.status(500).json({success: false, message: "Server error"}); 
     }
 }
